lambda/analyze/requirements.txt
google-genai>=1.0
psycopg2-binary>=2.9
boto3>=1.34
Pillow>=10.0
